update_tps_20190621
update_tps_20190621

+++ 10-common/include/tp/eventoutcncode.h
@@ -548,7 +548,7 @@ _ev_end
   *[消息方向
   *   cn  = > cnmpctrl 
   */_event ( ev_tppCfgSetMTUCmd )
-   _body( s32          ,1 )
+   _body( u16          ,1 )
    _ev_end
 
    /***********************<< 配置音质优先命令 >>************************
